backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
pillow
numpy
opencv-python
